fix: share markdown code block renderer
